requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
db-dtypes==1.0.5
Django==4.1.6
django-bootstrap-v5==1.0.11
django-environ==0.9.0
django-storages==1.13.2
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
google-cloud-datacatalog==3.11.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
idna==3.4
numpy==1.24.1
packaging==23.0
pandas==1.5.3
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.5
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14

pip~=22.3.1
wheel~=0.38.4
setuptools~=65.6.3
